+++ relays/bin-substrate/src/chains/rialto_messages_to_millau.rs
@@ -25,7 +25,9 @@ use sp_core::{Bytes, Pair};
 use bp_messages::MessageNonce;
 use bridge_runtime_common::messages::target::FromBridgedChainMessagesProof;
 use frame_support::weights::Weight;
-use messages_relay::{message_lane::MessageLane, relay_strategy::DefaultRelayStrategy};
+use messages_relay::{
+	message_lane::MessageLane, relay_strategy::altruistic_strategy::AltruisticStrategy,
+};
 use relay_millau_client::{
 	HeaderId as MillauHeaderId, Millau, SigningParams as MillauSigningParams,
 };
@@ -50,7 +52,7 @@ pub type MessageLaneRialtoMessagesToMillau = SubstrateMessageLaneToSubstrate<
 	RialtoSigningParams,
 	Millau,
 	MillauSigningParams,
-	DefaultRelayStrategy,
+	AltruisticStrategy,
 >;
 
 #[derive(Clone)]
