+++ relays/bin-substrate/src/chains/wococo_messages_to_rococo.rs
@@ -24,7 +24,9 @@ use sp_core::{Bytes, Pair};
 use bp_messages::MessageNonce;
 use bridge_runtime_common::messages::target::FromBridgedChainMessagesProof;
 use frame_support::weights::Weight;
-use messages_relay::{message_lane::MessageLane, relay_strategy::DefaultRelayStrategy};
+use messages_relay::{
+	message_lane::MessageLane, relay_strategy::altruistic_strategy::AltruisticStrategy,
+};
 use relay_rococo_client::{
 	HeaderId as RococoHeaderId, Rococo, SigningParams as RococoSigningParams,
 };
@@ -49,7 +51,7 @@ pub type MessageLaneWococoMessagesToRococo = SubstrateMessageLaneToSubstrate<
 	WococoSigningParams,
 	Rococo,
 	RococoSigningParams,
-	DefaultRelayStrategy,
+	AltruisticStrategy,
 >;
 
 #[derive(Clone)]
