+++ relays/lib-substrate-relay/src/messages_target.rs
@@ -437,7 +437,7 @@ fn compute_prepaid_messages_refund<P: SubstrateMessageLane>(
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use messages_relay::relay_strategy::DefaultRelayStrategy;
+	use messages_relay::relay_strategy::AltruisticStrategy;
 	use relay_rococo_client::{Rococo, SigningParams as RococoSigningParams};
 	use relay_wococo_client::{SigningParams as WococoSigningParams, Wococo};
 
@@ -450,7 +450,7 @@ mod tests {
 			RococoSigningParams,
 			Wococo,
 			WococoSigningParams,
-			DefaultRelayStrategy,
+			AltruisticStrategy,
 		>;
 
 		const OUTBOUND_LANE_MESSAGE_DETAILS_METHOD: &'static str = "";
