+++ relays/messages/src/message_race_delivery.rs
@@ -515,17 +515,18 @@ where
 			self.strategy.maximal_available_source_queue_index(race_state)?;
 		let previous_total_dispatch_weight = self.total_queued_dispatch_weight();
 		let source_queue = self.strategy.source_queue();
-		let range_end = select_nonces_for_delivery_transaction(
-			relayer_mode,
-			max_nonces,
+
+		let reference = RelayReference {
+			max_messages_in_this_batch: max_nonces,
 			max_messages_weight_in_single_batch,
 			max_messages_size_in_single_batch,
-			lane_source_client.clone(),
-			lane_target_client.clone(),
-			source_queue,
-			0..maximal_source_queue_index + 1,
-		)
-		.await?;
+			lane_source_client: lane_source_client.clone(),
+			lane_target_client: lane_target_client.clone(),
+			nonces_queue: source_queue.clone(),
+			nonces_queue_range: 0..maximal_source_queue_index + 1,
+		};
+
+		let range_end = P::RelayerStrategy::decide(reference).await?;
 
 		let range_begin = source_queue[0].1.begin();
 		let selected_nonces = range_begin..=range_end;
@@ -538,176 +539,6 @@ where
 			selected_nonces,
 			MessageProofParameters { outbound_state_proof_required, dispatch_weight },
 		))
-	}
-}
-
-/// From given set of source nonces, that are ready to be delivered, select nonces
-/// to fit into single delivery transaction.
-///
-/// The function returns last nonce that must be delivered to the target chain.
-#[allow(clippy::too_many_arguments)]
-async fn select_nonces_for_delivery_transaction<P: MessageLane>(
-	relayer_mode: RelayerMode,
-	max_messages_in_this_batch: MessageNonce,
-	max_messages_weight_in_single_batch: Weight,
-	max_messages_size_in_single_batch: u32,
-	lane_source_client: impl MessageLaneSourceClient<P>,
-	lane_target_client: impl MessageLaneTargetClient<P>,
-	nonces_queue: &SourceRangesQueue<
-		P::SourceHeaderHash,
-		P::SourceHeaderNumber,
-		MessageDetailsMap<P::SourceChainBalance>,
-	>,
-	nonces_queue_range: Range<usize>,
-) -> Option<MessageNonce> {
-	let mut hard_selected_count = 0;
-	let mut soft_selected_count = 0;
-
-	let mut selected_weight: Weight = 0;
-	let mut selected_unpaid_weight: Weight = 0;
-	let mut selected_prepaid_nonces = 0;
-	let mut selected_size: u32 = 0;
-	let mut selected_count: MessageNonce = 0;
-	let mut selected_reward = P::SourceChainBalance::zero();
-	let mut selected_cost = P::SourceChainBalance::zero();
-
-	let hard_selected_begin_nonce = nonces_queue[nonces_queue_range.start].1.begin();
-
-	let all_ready_nonces = nonces_queue
-		.range(nonces_queue_range.clone())
-		.flat_map(|(_, ready_nonces)| ready_nonces.iter())
-		.enumerate();
-	for (index, (nonce, details)) in all_ready_nonces {
-		// Since we (hopefully) have some reserves in `max_messages_weight_in_single_batch`
-		// and `max_messages_size_in_single_batch`, we may still try to submit transaction
-		// with single message if message overflows these limits. The worst case would be if
-		// transaction will be rejected by the target runtime, but at least we have tried.
-
-		// limit messages in the batch by weight
-		let new_selected_weight = match selected_weight.checked_add(details.dispatch_weight) {
-			Some(new_selected_weight)
-				if new_selected_weight <= max_messages_weight_in_single_batch =>
-				new_selected_weight,
-			new_selected_weight if selected_count == 0 => {
-				log::warn!(
-					target: "bridge",
-					"Going to submit message delivery transaction with declared dispatch \
-					weight {:?} that overflows maximal configured weight {}",
-					new_selected_weight,
-					max_messages_weight_in_single_batch,
-				);
-				new_selected_weight.unwrap_or(Weight::MAX)
-			},
-			_ => break,
-		};
-
-		// limit messages in the batch by size
-		let new_selected_size = match selected_size.checked_add(details.size) {
-			Some(new_selected_size) if new_selected_size <= max_messages_size_in_single_batch =>
-				new_selected_size,
-			new_selected_size if selected_count == 0 => {
-				log::warn!(
-					target: "bridge",
-					"Going to submit message delivery transaction with message \
-					size {:?} that overflows maximal configured size {}",
-					new_selected_size,
-					max_messages_size_in_single_batch,
-				);
-				new_selected_size.unwrap_or(u32::MAX)
-			},
-			_ => break,
-		};
-
-		// limit number of messages in the batch
-		let new_selected_count = selected_count + 1;
-		if new_selected_count > max_messages_in_this_batch {
-			break
-		}
-
-		// If dispatch fee has been paid at the source chain, it means that it is **relayer** who's
-		// paying for dispatch at the target chain AND reward must cover this dispatch fee.
-		//
-		// If dispatch fee is paid at the target chain, it means that it'll be withdrawn from the
-		// dispatch origin account AND reward is not covering this fee.
-		//
-		// So in the latter case we're not adding the dispatch weight to the delivery transaction
-		// weight.
-		let mut new_selected_prepaid_nonces = selected_prepaid_nonces;
-		let new_selected_unpaid_weight = match details.dispatch_fee_payment {
-			DispatchFeePayment::AtSourceChain => {
-				new_selected_prepaid_nonces += 1;
-				selected_unpaid_weight.saturating_add(details.dispatch_weight)
-			},
-			DispatchFeePayment::AtTargetChain => selected_unpaid_weight,
-		};
-
-		let reference = RelayReference {
-			relayer_mode,
-			lane_source_client: lane_source_client.clone(),
-			lane_target_client: lane_target_client.clone(),
-			hard_selected_begin_nonce,
-			new_selected_prepaid_nonces,
-			new_selected_unpaid_weight,
-			new_selected_size,
-			ready_nonces_index: index,
-			ready_nonce: *nonce,
-			ready_details: details.clone(),
-		};
-		let decide = P::RelayerStrategy::decide(reference).await?;
-		if decide.participate {
-			soft_selected_count = index + 1;
-			if let Some(total_reward) = decide.total_reward {
-				selected_reward = total_reward;
-			}
-			if let Some(total_cost) = decide.total_cost {
-				selected_cost = total_cost;
-			}
-		}
-
-		hard_selected_count = index + 1;
-		selected_weight = new_selected_weight;
-		selected_unpaid_weight = new_selected_unpaid_weight;
-		selected_prepaid_nonces = new_selected_prepaid_nonces;
-		selected_size = new_selected_size;
-		selected_count = new_selected_count;
-	}
-
-	if hard_selected_count != soft_selected_count {
-		let hard_selected_end_nonce =
-			hard_selected_begin_nonce + hard_selected_count as MessageNonce - 1;
-		let soft_selected_begin_nonce = hard_selected_begin_nonce;
-		let soft_selected_end_nonce =
-			soft_selected_begin_nonce + soft_selected_count as MessageNonce - 1;
-		log::warn!(
-			target: "bridge",
-			"Relayer may deliver nonces [{:?}; {:?}], but because of its strategy ({:?}) it has selected \
-			nonces [{:?}; {:?}].",
-			hard_selected_begin_nonce,
-			hard_selected_end_nonce,
-			relayer_mode,
-			soft_selected_begin_nonce,
-			soft_selected_end_nonce,
-		);
-
-		hard_selected_count = soft_selected_count;
-	}
-
-	if hard_selected_count != 0 {
-		if relayer_mode != RelayerMode::Altruistic {
-			log::trace!(
-				target: "bridge",
-				"Expected reward from delivering nonces [{:?}; {:?}] is: {:?} - {:?} = {:?}",
-				hard_selected_begin_nonce,
-				hard_selected_begin_nonce + hard_selected_count as MessageNonce - 1,
-				selected_reward,
-				selected_cost,
-				selected_reward - selected_cost,
-			);
-		}
-
-		Some(hard_selected_begin_nonce + hard_selected_count as MessageNonce - 1)
-	} else {
-		None
 	}
 }
